pyproject.toml
[project]
name = "ytune"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Nurahfezan Nordin", email = "nurahfezanbinnordin@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.12.5",
    "python-mpv>=1.0.8",
    "textual>=8.1.1",
    "yt-dlp>=2026.2.21",
    "ytmusicapi>=1.11.5",
]

[project.scripts]
ytune = "ytune:main"

[build-system]
requires = ["uv_build>=0.10.9,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pixi.dependencies]
python = "3.12.*"
mpv = ">=0.39.0"

[tool.pixi.tasks]
ytune = { cmd = "ytune", env = { LC_ALL = "C", LC_NUMERIC = "C" } }
test = "pytest"

[tool.pixi.environments]
default = { solve-group = "default" }
dev = { features = ["dev"], solve-group = "default" }

[tool.pixi.pypi-dependencies]
ytune = { path = ".", editable = true }

[tool.pixi.feature.dev.dependencies]
pytest = "*"
pytest-asyncio = "*"
pytest-mock = "*"
